2 add dynamic offset for the door distance calibration (#4)
* feat: Add 'Door Distance Offset' sensor and update template text

This commit adds a new sensor called 'Door Distance Offset' and updates the template text to reflect the changes. The sensor has a range from -0.9 to 0 with a step of 0.1 and is measured in meters. The initial value is set to -0.5. The template text is updated to display the new value of 'Door Distance Offset' when it changes.

* chore: Update door distance offset calculation

* feat: Add 'Door Distance Offset' sensor and update template text

---------

Co-authored-by: Thom402 <thomas.abel234@gamil.com>